final_poly stuff moved
